nsu-audit-engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
PyJWT>=2.8.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
google-auth>=2.22.0
requests>=2.31.0
httpx>=0.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
